Move MainActivity and related fragment to package dashboard
