+++ app/src/main/AndroidManifest.xml
@@ -22,7 +22,7 @@
         </activity>
 
         <activity
-            android:name=".ui.MainActivity"
+            android:name=".ui.dashboard.MainActivity"
             android:theme="@style/Theme.AyoPintar"
             android:exported="true">
         </activity>
