+++ app/src/main/java/com/example/ayopintar/ui/dashboard/home/HomeFragment.kt
@@ -1,4 +1,4 @@
-package com.example.ayopintar.ui.profesi
+package com.example.ayopintar.ui.dashboard.home
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -8,14 +8,12 @@ import android.view.ViewGroup
 import com.example.ayopintar.R
 
 
-class ProfesiFragment : Fragment() {
-
+class HomeFragment : Fragment() {
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_profesi, container, false)
+        return inflater.inflate(R.layout.fragment_home, container, false)
     }
-
 }
