+++ app/src/main/java/com/example/ayopintar/ui/dashboard/profesi/ProfesiFragment.kt
@@ -1,4 +1,4 @@
-package com.example.ayopintar.ui.tipsbelajar
+package com.example.ayopintar.ui.dashboard.profesi
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -7,15 +7,15 @@ import android.view.View
 import android.view.ViewGroup
 import com.example.ayopintar.R
 
-class TipsBelajarFragment : Fragment() {
+
+class ProfesiFragment : Fragment() {
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_tips_belajar, container, false)
+        return inflater.inflate(R.layout.fragment_profesi, container, false)
     }
-
 
 }
