+++ app/src/main/java/com/example/ayopintar/ui/dashboard/tipsbelajar/TipsBelajarFragment.kt
@@ -1,4 +1,4 @@
-package com.example.ayopintar.ui.home
+package com.example.ayopintar.ui.dashboard.tipsbelajar
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -7,13 +7,15 @@ import android.view.View
 import android.view.ViewGroup
 import com.example.ayopintar.R
 
+class TipsBelajarFragment : Fragment() {
 
-class HomeFragment : Fragment() {
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_home, container, false)
+        return inflater.inflate(R.layout.fragment_tips_belajar, container, false)
     }
+
+
 }
