+++ app/src/main/res/layout/fragment_profesi.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.profesi.ProfesiFragment">
+    tools:context=".ui.dashboard.profesi.ProfesiFragment">
 
 
     <TextView
