+++ app/src/main/res/layout/fragment_tips_belajar.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.tipsbelajar.TipsBelajarFragment">
+    tools:context=".ui.dashboard.tipsbelajar.TipsBelajarFragment">
 
     <TextView
         android:layout_width="match_parent"
